correção das buscas

+++ Model/Entity/ResultEntity.cs
@@ -2,7 +2,7 @@
 {
     public class ResultEntity
     {
-        public int Result { get; set; }
+        public long Result { get; set; }
         public string ErrorMessage { get; set; }
 
         public bool IsSuccess { get { return ErrorMessage.Equals("Success"); } }
